Add ProcessSample interface
