+++ lib/src/Base/Stat/CMakeLists.txt
@@ -25,6 +25,7 @@ ot_add_source_file (Null.cxx)
 ot_add_source_file (Sample.cxx)
 ot_add_source_file (SampleImplementation.cxx)
 ot_add_source_file (ProcessSample.cxx)
+ot_add_source_file (ProcessSampleImplementation.cxx)
 ot_add_source_file (RandomGenerator.cxx)
 ot_add_source_file (RandomGeneratorState.cxx)
 ot_add_source_file (ReverseHaltonSequence.cxx)
@@ -93,6 +94,7 @@ ot_install_header_file (Full.hxx)
 ot_install_header_file (SampleImplementation.hxx)
 ot_install_header_file (Sample.hxx)
 ot_install_header_file (ProcessSample.hxx)
+ot_install_header_file (ProcessSampleImplementation.hxx)
 ot_install_header_file (OTStat.hxx)
 ot_install_header_file (Compact.hxx)
 ot_install_header_file (LowDiscrepancySequenceImplementation.hxx)
