+++ lib/src/Base/Stat/ProcessSample.cxx
@@ -20,29 +20,13 @@
  */
 
 #include "openturns/ProcessSample.hxx"
-#include "openturns/PersistentObjectFactory.hxx"
-#include "openturns/Exception.hxx"
-#include "openturns/ResourceMap.hxx"
-#include "openturns/Drawable.hxx"
-#include "openturns/Description.hxx"
-#include "openturns/Log.hxx"
 
 BEGIN_NAMESPACE_OPENTURNS
 
-TEMPLATE_CLASSNAMEINIT(PersistentCollection<ProcessSample>)
-
-TEMPLATE_CLASSNAMEINIT(PersistentCollection<Field>)
-
-static const Factory<PersistentCollection<Field> > Factory_PersistentCollection_Field;
-
 CLASSNAMEINIT(ProcessSample)
 
-static const Factory<ProcessSample> Factory_ProcessSample;
-
 ProcessSample::ProcessSample()
-  : PersistentObject()
-  , mesh_()
-  , data_(0)
+  : TypedInterfaceObject<ProcessSampleImplementation>(new ProcessSampleImplementation)
 {
   // Nothing to do
 }
@@ -50,9 +34,7 @@ ProcessSample::ProcessSample()
 /* Default constructor */
 ProcessSample::ProcessSample(const UnsignedInteger size,
                              const Field & field)
-  : PersistentObject()
-  , mesh_(field.getMesh())
-  , data_(SampleCollection(size, field.getValues()))
+  : TypedInterfaceObject<ProcessSampleImplementation>(new ProcessSampleImplementation(size, field))
 {
   // Nothing to do
 }
@@ -60,137 +42,121 @@ ProcessSample::ProcessSample(const UnsignedInteger size,
 ProcessSample::ProcessSample(const Mesh & mesh,
                              const UnsignedInteger size,
                              const UnsignedInteger dimension)
-  : PersistentObject()
-  , mesh_(mesh)
-  , data_(SampleCollection(size, Sample(mesh.getVerticesNumber(), dimension)))
+  : TypedInterfaceObject<ProcessSampleImplementation>(new ProcessSampleImplementation(mesh, size, dimension))
 {
   // Nothing to do
 }
 
-/* Virtual constructor */
-ProcessSample * ProcessSample::clone() const
+
+/* Parameters constructor */
+ProcessSample::ProcessSample(const ProcessSampleImplementation & implementation)
+  : TypedInterfaceObject<ProcessSampleImplementation>(implementation.clone())
 {
-  return new ProcessSample(*this);
+  // Nothing to do
+}
+
+/* Constructor from implementation */
+ProcessSample::ProcessSample(const Implementation & p_implementation)
+  : TypedInterfaceObject<ProcessSampleImplementation>(p_implementation)
+{
+  // Nothing to do
+}
+
+/* Constructor from implementation pointer */
+ProcessSample::ProcessSample(ProcessSampleImplementation * p_implementation)
+  : TypedInterfaceObject<ProcessSampleImplementation>(p_implementation)
+{
+  // Nothing to do
 }
 
 /* String converter */
 String ProcessSample::__repr__() const
 {
-  OSS oss(true);
-  oss << "class=" << ProcessSample::GetClassName()
-      << " mesh=" << mesh_
-      << " values=" << data_;
-  return oss;
+  return getImplementation()->__repr__();
 }
 
 String ProcessSample::__str__(const String & offset) const
 {
-  OSS oss(false);
-  oss << offset << "[";
-  String separator("");
-  for (UnsignedInteger i = 0; i < data_.getSize(); ++i, separator = "\n") oss << separator << offset << "field " << i << ":\n" << getField(i).__str__(offset);
-  oss << "]";
-  return oss;
+  return getImplementation()->__str__(offset);
 }
 
 /* Partial copy constructor */
 void ProcessSample::add(const Field & field)
 {
-  if (getSize() == 0)
-  {
-    data_.add(field.getValues());
-    mesh_ = field.getMesh();
-  }
-  else if ((data_[0].getDimension() == field.getOutputDimension()) && (mesh_ == field.getMesh())) data_.add(field.getValues());
-  else throw InvalidArgumentException(HERE) << "Error: could not add the field. Either its dimenson or its mesh are incompatible.";
+  copyOnWrite();
+  getImplementation()->add(field);
 }
 
 void ProcessSample::add(const Sample & values)
 {
-  if (values.getSize() != mesh_.getVerticesNumber()) throw InvalidArgumentException(HERE) << "Error: could not add the values. Their size=" << values.getSize() << " does not match the number of vertices=" << mesh_.getVerticesNumber() << " of the mesh.";
-  if ((getSize() > 0) && (data_[0].getDimension() != values.getDimension())) throw InvalidArgumentException(HERE) << "Error: could not add the values. Their dimension=" << values.getDimension() << " does not match the process sample dimension=" << data_[0].getDimension();
-  data_.add(values);
+  copyOnWrite();
+  getImplementation()->add(values);
 }
 
 
 /* Operators accessors */
 Field ProcessSample::getField(const UnsignedInteger index) const
 {
-  if (index >= data_.getSize()) throw OutOfBoundException(HERE)  << " Error - index should be between 0 and " << data_.getSize() - 1;
-  return Field(mesh_, data_[index]);
+  return getImplementation()->getField(index);
 }
 
 void ProcessSample::setField(const Field & field,
                              const UnsignedInteger index)
 {
-  if (index >= data_.getSize()) throw OutOfBoundException(HERE)  << " Error - index should be between 0 and " << data_.getSize() - 1;
-  if (field.getOutputDimension() != data_[0].getDimension()) throw InvalidArgumentException(HERE) << "Error: expected a field of dimension=" << data_[0].getDimension() << ", got a field of dimension=" << field.getOutputDimension();
-  data_[index] = field.getValues();
+  copyOnWrite();
+  getImplementation()->setField(field, index);
 }
 
 Sample & ProcessSample::operator[] (const UnsignedInteger index)
 {
-  if (index >= data_.getSize()) throw OutOfBoundException(HERE)  << " Error - index should be between 0 and " << data_.getSize() - 1;
-  return data_[index];
+  copyOnWrite();
+  return getImplementation()->operator[](index);
 }
 
 const Sample & ProcessSample::operator[] (const UnsignedInteger index) const
 {
-  if (index >= data_.getSize()) throw OutOfBoundException(HERE)  << " Error - index should be between 0 and " << data_.getSize() - 1;
-  return data_[index];
+  return getImplementation()->operator[](index);
 }
 
 /* Time grid accessors */
 RegularGrid ProcessSample::getTimeGrid() const
 {
-  return RegularGrid(mesh_);
+  return getImplementation()->getTimeGrid();
 }
 
 /* Mesh accessors */
 Mesh ProcessSample::getMesh() const
 {
-  return mesh_;
+  return getImplementation()->getMesh();
 }
 
 /* Dimension accessors */
 UnsignedInteger ProcessSample::getDimension() const
 {
-  if (data_.getSize() == 0) return 0;
-  return data_[0].getDimension();
+  return getImplementation()->getDimension();
 }
 
 /* Dimension accessors */
 UnsignedInteger ProcessSample::getSize() const
 {
-  return data_.getSize();
+  return getImplementation()->getSize();
 }
 
 Field ProcessSample::computeMean() const
 {
-  const UnsignedInteger size = getSize();
-  if (size == 0) return Field();
-  if (size == 1) return Field(mesh_, data_[0]);
-  Sample meanValues(data_[0]);
-  for (UnsignedInteger i = 1; i < size; ++i) meanValues += data_[i];
-  meanValues *= Point(getDimension(), 1.0 / size);
-  return Field(mesh_, meanValues);
+  return getImplementation()->computeMean();
 }
 
 /* Compute the sample of spatial means of each field */
 Sample ProcessSample::computeTemporalMean() const
 {
-  if (!mesh_.isRegular() || (mesh_.getDimension() != 1)) throw InvalidArgumentException(HERE) << "Error: the temporal mean is defined only when the mesh is regular and of dimension 1.";
-  return computeSpatialMean();
+  return getImplementation()->computeTemporalMean();
 }
 
 /* Compute the sample of spatial means of each field */
 Sample ProcessSample::computeSpatialMean() const
 {
-  const UnsignedInteger size = getSize();
-  const UnsignedInteger dimension = getDimension();
-  Sample result(size, dimension);
-  for (UnsignedInteger i = 0; i < size; ++i) result[i] = data_[i].computeMean();
-  return result;
+  return getImplementation()->computeSpatialMean();
 }
 
 /*
@@ -198,76 +164,26 @@ Sample ProcessSample::computeSpatialMean() const
  */
 Field ProcessSample::computeQuantilePerComponent(const Scalar prob) const
 {
-  const UnsignedInteger size = getSize();
-  if (size == 0) return Field();
-  if (size == 1) return Field(mesh_, data_[0]);
-  // This initialization set the correct time grid into result
-  const UnsignedInteger dimension = data_[0].getDimension();
-  const UnsignedInteger length = data_[0].getSize();
-  Sample result(length, dimension);
-  // Loop over the location indices
-  for (UnsignedInteger i = 0; i < length; ++i)
-  {
-    Sample dataI(size, dimension);
-    for (UnsignedInteger j = 0; j < size; ++j)
-      dataI[j] = data_[j][i];
-    result[i] = dataI.computeQuantilePerComponent(prob);
-  }
-  return Field(mesh_, result);
+  return getImplementation()->computeQuantilePerComponent(prob);
 }
 
 /* Get the i-th marginal process sample */
 ProcessSample ProcessSample::getMarginal(const UnsignedInteger index) const
 {
-  const UnsignedInteger size = data_.getSize();
-  ProcessSample result(mesh_, size, 1);
-  for (UnsignedInteger i = 0; i < size; ++i) result[i] = data_[i].getMarginal(index);
-  return result;
+  return getImplementation()->getMarginal(index);
 }
 
 /* Get the marginal field corresponding to indices dimensions */
 ProcessSample ProcessSample::getMarginal(const Indices & indices) const
 {
-  const UnsignedInteger size = data_.getSize();
-  ProcessSample result(mesh_, size, indices.getSize());
-  for (UnsignedInteger i = 0; i < size; ++i) result[i] = data_[i].getMarginal(indices);
-  return result;
+  return getImplementation()->getMarginal(indices);
 }
 
 /* Draw a marginal of the ProcessSample, ie the collection of all the Field marginals */
 Graph ProcessSample::drawMarginal(const UnsignedInteger index) const
 {
-  if (mesh_.getDimension() != 1) throw NotDefinedException(HERE) << "Error: cannot draw a marginal sample if the mesh is of dimension greater than one. Here dimension=" << mesh_.getDimension();
-  if (index > getDimension() - 1 ) throw InvalidArgumentException(HERE) << "Error : indice should be in {0,...," << getDimension() - 1 << "}";
-
-  // Discretization of the x axis
-  const String title(OSS() << getName() << " - " << index << " marginal" );
-  Graph graph(title, "Time", "Values", true, "topright");
-  const UnsignedInteger size = data_.getSize();
-  const Description colors(Drawable::BuildDefaultPalette(size));
-  for (UnsignedInteger i = 0; i < size; ++i)
-  {
-    Drawable drawable(Field(mesh_, data_[i]).drawMarginal(index).getDrawable(0));
-    drawable.setColor(colors[i]);
-    graph.add(drawable);
-  }
-  return graph;
+  return getImplementation()->drawMarginal(index);
 }
 
-/* Method save() stores the object through the StorageManager */
-void ProcessSample::save(Advocate & adv) const
-{
-  PersistentObject::save(adv);
-  adv.saveAttribute( "mesh_", mesh_);
-  adv.saveAttribute( "data_", data_ );
-}
-
-/* Method load() reloads the object from the StorageManager */
-void ProcessSample::load(Advocate & adv)
-{
-  PersistentObject::load(adv);
-  adv.loadAttribute( "mesh_", mesh_);
-  adv.loadAttribute( "data_", data_ );
-}
 
 END_NAMESPACE_OPENTURNS
