+++ lib/test/t_ARMA_std.expout
@@ -8,4 +8,4 @@ process2 = class= ARMA timeGrid=class=RegularGrid name=Unnamed start=0 step=0.1 
 process3 = class= ARMA timeGrid=class=RegularGrid name=Unnamed start=0 step=0.1 n=11 coefficients AR=class=ARMACoefficients, shift=0, value=class=SquareMatrix dimension=2 implementation=class=MatrixImplementation name=Unnamed rows=2 columns=2 values=[0.2,0.3,0.7,0.4] coefficients MA=class=ARMACoefficients, shift=0, value=class=SquareMatrix dimension=2 implementation=class=MatrixImplementation name=Unnamed rows=2 columns=2 values=[0.1,0,0,0.5] noiseDistribution= class=ComposedDistribution name=ComposedDistribution dimension=2 copula=class=IndependentCopula name=IndependentCopula dimension=2 marginal[0]=class=Normal name=Normal dimension=1 mean=class=Point name=Unnamed dimension=1 values=[0] sigma=class=Point name=Unnamed dimension=1 values=[0.01] correlationMatrix=class=CorrelationMatrix dimension=1 implementation=class=MatrixImplementation name=Unnamed rows=1 columns=1 values=[1] marginal[1]=class=Normal name=Normal dimension=1 mean=class=Point name=Unnamed dimension=1 values=[0] sigma=class=Point name=Unnamed dimension=1 values=[0.02] correlationMatrix=class=CorrelationMatrix dimension=1 implementation=class=MatrixImplementation name=Unnamed rows=1 columns=1 values=[1] state= class= ARMAState x= class=Sample name=Unnamed implementation=class=SampleImplementation name=Unnamed size=1 dimension=2 data=[[0.629877,0.135276]] epsilon= class=Sample name=Unnamed implementation=class=SampleImplementation name=Unnamed size=1 dimension=2 data=[[0.882805,0.0325028]]
 One realization=class=Field name=Unnamed description=[t,X0,X1] implementation=class=FieldImplementation name=Unnamed mesh=class=Mesh name=Unnamed dimension=1 vertices=class=Sample name=Unnamed implementation=class=SampleImplementation name=Unnamed size=11 dimension=1 description=[t] data=[[0],[0.1],[0.2],[0.3],[0.4],[0.5],[0.6],[0.7],[0.8],[0.9],[1]] simplices=[[0,1],[1,2],[2,3],[3,4],[4,5],[5,6],[6,7],[7,8],[8,9],[9,10]] values=class=Sample name=Unnamed implementation=class=SampleImplementation name=Unnamed size=11 dimension=2 description=[X0,X1] data=[[0.010204,-0.00744086],[-0.00287281,-0.0293307],[0.0190836,-0.00416576],[0.000428186,-0.0133922],[0.0226688,-0.0154754],[0.0156377,-0.00180096],[-0.00152644,0.0220058],[-0.0290278,-0.000953817],[0.00341936,-0.0137703],[0.00530122,-0.00738418],[-0.0161936,0.0111398]]
 One future=class=TimeSeries name=Unnamed derived from=class=FieldImplementation name=Unnamed mesh=class=Mesh name=Unnamed dimension=1 vertices=class=Sample name=Unnamed implementation=class=SampleImplementation name=Unnamed size=4 dimension=1 description=[t] data=[[1.1],[1.2],[1.3],[1.4]] simplices=[[0,1],[1,2],[2,3]] values=class=Sample name=Unnamed implementation=class=SampleImplementation name=Unnamed size=4 dimension=2 data=[[-0.00486985,0.0088404],[-0.0213829,0.0140495],[-0.0034992,0.00524708],[0.0067867,-0.00506869]] start=1.1 timeStep=0.1 n=4
-Some futures=class=ProcessSample mesh=class=Mesh name=Unnamed dimension=1 vertices=class=Sample name=Unnamed implementation=class=SampleImplementation name=Unnamed size=4 dimension=1 description=[t] data=[[1.1],[1.2],[1.3],[1.4]] simplices=[[0,1],[1,2],[2,3]] values=[class=Sample name=Unnamed implementation=class=SampleImplementation name=Unnamed size=4 dimension=2 data=[[-1.82364e-05,0.0156951],[-0.0150433,-0.0192013],[0.0304955,0.00743803],[-0.00879671,-0.0355234]],class=Sample name=Unnamed implementation=class=SampleImplementation name=Unnamed size=4 dimension=2 data=[[-0.0129397,-0.00706455],[0.013426,-0.0179524],[-0.00898944,0.00403051],[-0.00677231,0.00868747]],class=Sample name=Unnamed implementation=class=SampleImplementation name=Unnamed size=4 dimension=2 data=[[-0.00500545,-0.0157516],[0.0236257,0.0301551],[-0.0128599,-0.00625936],[-0.00239186,0.000610274]]]
+Some futures=class=ProcessSampleImplementation mesh=class=Mesh name=Unnamed dimension=1 vertices=class=Sample name=Unnamed implementation=class=SampleImplementation name=Unnamed size=4 dimension=1 description=[t] data=[[1.1],[1.2],[1.3],[1.4]] simplices=[[0,1],[1,2],[2,3]] values=[class=Sample name=Unnamed implementation=class=SampleImplementation name=Unnamed size=4 dimension=2 data=[[-1.82364e-05,0.0156951],[-0.0150433,-0.0192013],[0.0304955,0.00743803],[-0.00879671,-0.0355234]],class=Sample name=Unnamed implementation=class=SampleImplementation name=Unnamed size=4 dimension=2 data=[[-0.0129397,-0.00706455],[0.013426,-0.0179524],[-0.00898944,0.00403051],[-0.00677231,0.00868747]],class=Sample name=Unnamed implementation=class=SampleImplementation name=Unnamed size=4 dimension=2 data=[[-0.00500545,-0.0157516],[0.0236257,0.0301551],[-0.0128599,-0.00625936],[-0.00239186,0.000610274]]]
