+++ lib/test/t_ProcessSample_std.expout
@@ -1,8 +1,8 @@
 Default constructor
-psample0=class=ProcessSample mesh=class=Mesh name=Unnamed dimension=1 vertices=class=Sample name=Unnamed implementation=class=SampleImplementation name=Unnamed size=1 dimension=1 data=[[0]] simplices=[] values=[]
+psample0=class=ProcessSampleImplementation mesh=class=Mesh name=Unnamed dimension=1 vertices=class=Sample name=Unnamed implementation=class=SampleImplementation name=Unnamed size=1 dimension=1 data=[[0]] simplices=[] values=[]
 Constructor based on size, dimension and timeGrid
-psample1=class=ProcessSample mesh=class=Mesh name=Unnamed dimension=1 vertices=class=Sample name=Unnamed implementation=class=SampleImplementation name=Unnamed size=6 dimension=1 description=[t] data=[[0],[0.2],[0.4],[0.6],[0.8],[1]] simplices=[[0,1],[1,2],[2,3],[3,4],[4,5]] values=[class=Sample name=Unnamed implementation=class=SampleImplementation name=Unnamed size=6 dimension=1 data=[[0],[0],[0],[0],[0],[0]],class=Sample name=Unnamed implementation=class=SampleImplementation name=Unnamed size=6 dimension=1 data=[[0],[0],[0],[0],[0],[0]],class=Sample name=Unnamed implementation=class=SampleImplementation name=Unnamed size=6 dimension=1 data=[[0],[0],[0],[0],[0],[0]],class=Sample name=Unnamed implementation=class=SampleImplementation name=Unnamed size=6 dimension=1 data=[[0],[0],[0],[0],[0],[0]]]
+psample1=class=ProcessSampleImplementation mesh=class=Mesh name=Unnamed dimension=1 vertices=class=Sample name=Unnamed implementation=class=SampleImplementation name=Unnamed size=6 dimension=1 description=[t] data=[[0],[0.2],[0.4],[0.6],[0.8],[1]] simplices=[[0,1],[1,2],[2,3],[3,4],[4,5]] values=[class=Sample name=Unnamed implementation=class=SampleImplementation name=Unnamed size=6 dimension=1 data=[[0],[0],[0],[0],[0],[0]],class=Sample name=Unnamed implementation=class=SampleImplementation name=Unnamed size=6 dimension=1 data=[[0],[0],[0],[0],[0],[0]],class=Sample name=Unnamed implementation=class=SampleImplementation name=Unnamed size=6 dimension=1 data=[[0],[0],[0],[0],[0],[0]],class=Sample name=Unnamed implementation=class=SampleImplementation name=Unnamed size=6 dimension=1 data=[[0],[0],[0],[0],[0],[0]]]
 changing psample1[0] with []
 psample1[0]=class=Sample name=Unnamed implementation=class=SampleImplementation name=Unnamed size=6 dimension=1 data=[[1],[2],[3],[4],[5],[6]]
 Constructor based on size / timeSerie
-psample2=class=ProcessSample mesh=class=Mesh name=Unnamed dimension=1 vertices=class=Sample name=Unnamed implementation=class=SampleImplementation name=Unnamed size=6 dimension=1 description=[t] data=[[0],[0.2],[0.4],[0.6],[0.8],[1]] simplices=[[0,1],[1,2],[2,3],[3,4],[4,5]] values=[class=Sample name=Unnamed implementation=class=SampleImplementation name=Unnamed size=6 dimension=1 data=[[1],[2],[3],[4],[5],[6]],class=Sample name=Unnamed implementation=class=SampleImplementation name=Unnamed size=6 dimension=1 data=[[1],[2],[3],[4],[5],[6]],class=Sample name=Unnamed implementation=class=SampleImplementation name=Unnamed size=6 dimension=1 data=[[1],[2],[3],[4],[5],[6]]]
+psample2=class=ProcessSampleImplementation mesh=class=Mesh name=Unnamed dimension=1 vertices=class=Sample name=Unnamed implementation=class=SampleImplementation name=Unnamed size=6 dimension=1 description=[t] data=[[0],[0.2],[0.4],[0.6],[0.8],[1]] simplices=[[0,1],[1,2],[2,3],[3,4],[4,5]] values=[class=Sample name=Unnamed implementation=class=SampleImplementation name=Unnamed size=6 dimension=1 data=[[1],[2],[3],[4],[5],[6]],class=Sample name=Unnamed implementation=class=SampleImplementation name=Unnamed size=6 dimension=1 data=[[1],[2],[3],[4],[5],[6]],class=Sample name=Unnamed implementation=class=SampleImplementation name=Unnamed size=6 dimension=1 data=[[1],[2],[3],[4],[5],[6]]]
