+++ lib/test/t_RandomWalk_std.expout
@@ -2,4 +2,4 @@ dist = class=Uniform name=Uniform dimension=1 a=-1 b=1
 process = class=Process implementation=class=RandomWalk, origin=class=Point name=Unnamed dimension=1 values=[0], distribution=class=Uniform name=Uniform dimension=1 a=-1 b=1
 timeSerie = class=TimeSeries name=Unnamed derived from=class=FieldImplementation name=Unnamed mesh=class=Mesh name=Unnamed dimension=1 vertices=class=Sample name=Unnamed implementation=class=SampleImplementation name=Unnamed size=11 dimension=1 description=[t] data=[[0],[0.1],[0.2],[0.3],[0.4],[0.5],[0.6],[0.7],[0.8],[0.9],[1]] simplices=[[0,1],[1,2],[2,3],[3,4],[4,5],[5,6],[6,7],[7,8],[8,9],[9,10]] values=class=Sample name=Unnamed implementation=class=SampleImplementation name=Unnamed size=11 dimension=1 description=[X0] data=[[0],[0.259753],[1.02536],[0.295916],[-0.639078],[-0.944964],[-0.00611811],[0.835241],[0.841321],[-0.0322665],[-0.446753]] start=0 timeStep=0.1 n=11
 One future=class=TimeSeries name=Unnamed derived from=class=FieldImplementation name=Unnamed mesh=class=Mesh name=Unnamed dimension=1 vertices=class=Sample name=Unnamed implementation=class=SampleImplementation name=Unnamed size=4 dimension=1 description=[t] data=[[1.1],[1.2],[1.3],[1.4]] simplices=[[0,1],[1,2],[2,3]] values=class=Sample name=Unnamed implementation=class=SampleImplementation name=Unnamed size=4 dimension=1 data=[[-0.0179887],[-0.251265],[-0.50373],[-0.0291936]] start=1.1 timeStep=0.1 n=4
-Some futures=class=ProcessSample mesh=class=Mesh name=Unnamed dimension=1 vertices=class=Sample name=Unnamed implementation=class=SampleImplementation name=Unnamed size=4 dimension=1 description=[t] data=[[1.1],[1.2],[1.3],[1.4]] simplices=[[0,1],[1,2],[2,3]] values=[class=Sample name=Unnamed implementation=class=SampleImplementation name=Unnamed size=4 dimension=1 data=[[0.320254],[-0.0897576],[0.767263],[1.40888]],class=Sample name=Unnamed implementation=class=SampleImplementation name=Unnamed size=4 dimension=1 data=[[-0.0776029],[0.578452],[0.298056],[1.20755]],class=Sample name=Unnamed implementation=class=SampleImplementation name=Unnamed size=4 dimension=1 data=[[-0.269522],[-0.905444],[-1.73387],[-1.41242]]]
+Some futures=class=ProcessSampleImplementation mesh=class=Mesh name=Unnamed dimension=1 vertices=class=Sample name=Unnamed implementation=class=SampleImplementation name=Unnamed size=4 dimension=1 description=[t] data=[[1.1],[1.2],[1.3],[1.4]] simplices=[[0,1],[1,2],[2,3]] values=[class=Sample name=Unnamed implementation=class=SampleImplementation name=Unnamed size=4 dimension=1 data=[[0.320254],[-0.0897576],[0.767263],[1.40888]],class=Sample name=Unnamed implementation=class=SampleImplementation name=Unnamed size=4 dimension=1 data=[[-0.0776029],[0.578452],[0.298056],[1.20755]],class=Sample name=Unnamed implementation=class=SampleImplementation name=Unnamed size=4 dimension=1 data=[[-0.269522],[-0.905444],[-1.73387],[-1.41242]]]
