+++ lib/test/t_WhiteNoise_std.expout
@@ -2,4 +2,4 @@ dist = class=Uniform name=Uniform dimension=1 a=-1 b=1
 process = class=Process implementation=class=WhiteNoise distribution=class=Uniform name=Uniform dimension=1 a=-1 b=1
 timeSerie = class=TimeSeries name=Unnamed derived from=class=FieldImplementation name=Unnamed mesh=class=Mesh name=Unnamed dimension=1 vertices=class=Sample name=Unnamed implementation=class=SampleImplementation name=Unnamed size=11 dimension=1 description=[t] data=[[0],[0.1],[0.2],[0.3],[0.4],[0.5],[0.6],[0.7],[0.8],[0.9],[1]] simplices=[[0,1],[1,2],[2,3],[3,4],[4,5],[5,6],[6,7],[7,8],[8,9],[9,10]] values=class=Sample name=Uniform implementation=class=SampleImplementation name=Uniform size=11 dimension=1 description=[X0] data=[[0.259753],[0.76561],[-0.729447],[-0.934994],[-0.305886],[0.938846],[0.841359],[0.0060803],[-0.873588],[-0.414486],[0.428764]] start=0 timeStep=0.1 n=11
 One future=class=TimeSeries name=Unnamed derived from=class=FieldImplementation name=Unnamed mesh=class=Mesh name=Unnamed dimension=1 vertices=class=Sample name=Unnamed implementation=class=SampleImplementation name=Unnamed size=4 dimension=1 description=[t] data=[[1.1],[1.2],[1.3],[1.4]] simplices=[[0,1],[1,2],[2,3]] values=class=Sample name=Uniform implementation=class=SampleImplementation name=Uniform size=4 dimension=1 description=[X0] data=[[-0.233276],[-0.252465],[0.474536],[0.767007]] start=1.1 timeStep=0.1 n=4
-Some futures=class=ProcessSample mesh=class=Mesh name=Unnamed dimension=1 vertices=class=Sample name=Unnamed implementation=class=SampleImplementation name=Unnamed size=4 dimension=1 description=[t] data=[[1.1],[1.2],[1.3],[1.4]] simplices=[[0,1],[1,2],[2,3]] values=[class=Sample name=Uniform implementation=class=SampleImplementation name=Uniform size=4 dimension=1 description=[X0] data=[[-0.410012],[0.857021],[0.641622],[0.36915]],class=Sample name=Uniform implementation=class=SampleImplementation name=Uniform size=4 dimension=1 description=[X0] data=[[0.656054],[-0.280396],[0.909492],[0.177231]],class=Sample name=Uniform implementation=class=SampleImplementation name=Uniform size=4 dimension=1 description=[X0] data=[[-0.635922],[-0.82843],[0.321455],[-0.579116]]]
+Some futures=class=ProcessSampleImplementation mesh=class=Mesh name=Unnamed dimension=1 vertices=class=Sample name=Unnamed implementation=class=SampleImplementation name=Unnamed size=4 dimension=1 description=[t] data=[[1.1],[1.2],[1.3],[1.4]] simplices=[[0,1],[1,2],[2,3]] values=[class=Sample name=Uniform implementation=class=SampleImplementation name=Uniform size=4 dimension=1 description=[X0] data=[[-0.410012],[0.857021],[0.641622],[0.36915]],class=Sample name=Uniform implementation=class=SampleImplementation name=Uniform size=4 dimension=1 description=[X0] data=[[0.656054],[-0.280396],[0.909492],[0.177231]],class=Sample name=Uniform implementation=class=SampleImplementation name=Uniform size=4 dimension=1 description=[X0] data=[[-0.635922],[-0.82843],[0.321455],[-0.579116]]]
